ByteArrayDiff now operational!
This puts in place a working version of ByteArrayDiff.  The granularity
of deltas generated could potentially be improved, especially for going
on the ledger.  However, it's now passing all tests.

Have also added a duplicate ObjectArrayDiff which can handngle generic
arrays.  Need to add test cases for this still, and perhaps attempt to
generalise the API a little.