Updated proxy design for JLedger
This is yet another evolution of the design.  It's much improved and is
workable.  However, my feeling is that we can go further.  Potentially
by requiring user-defined proxies to implement their own layout.

The other possibility is that we could extend Content.Blob to be fully
aware of what it's contents are.  This could be powerful in the sense
that the outermost proxy *is* the blob.