Working on BuildServer and diff
At this stage, the build server writes the entire new contents of the file into
the ledger.  No attempt is made to produce a minimal diff.  Somehow, this not
even that easy to do.  Potentially could use Byte.diff() within the Byte.Diff
write method's to minimise *all* writes.  This would further apply to integer
and short writes, etc.