Working on diffs within BuildServer
The key problem is that there is currently no straightforward mechanism to write
a "diff" into a given Primitive.ByteArrayLayout.  Basic issues:

(1) Cannot apply a diff into a primitive byte array.  Whilst for BuildServer
this is actually unnecessary, it would be necessary for other systems.

(2) Currently, writing a complete file saves the entire contents even if there
is a lot of overlap.  This could be resolved using the Byte.diff() algorithm at
the point of the write.

To fix (1) above, we actually need a proxy for a byte array so that we can
describe the concept of a partial write.  Otherwise, we have to just write bytes
individually and hope they all get merged backtogether in the diff.  It's not
ideal!