Working on Efficient Writes
Have been trying to figure out how best to generate efficient data
writes to blob's and ledgers.  Specific observations:

(1) Created Primitive.ByteArray as a byte[] proxy.  This doesn't get me
much at the moment, but potentially could allow for partial updates.
The problem is, even if it supported partial updates, I don't know how
they would be applied through e.g. Pair.setFirst(), etc.

(2) Have explored doing a diff in the Blob.writeBytes() and
Blob.replaceBytes() functions.  This can work, though care is needed to
ensure a diff is actually smaller than a complete replacement.  Also
making it efficient requires further development of the Byte.diff API to
allow for the unusual layout of a ByteArrayLedger.

Part of the difficulty here is deciding upon the "granularity" of diffs.
Do we want every single write to generate a diff, or do we want the user
to decide when diff's are generated.  The former provides for a pretty
simply API, whilst the latter is presumably more efficient in general.
That is, because the user presumably knows when a diff is likely to pay
off.

Further thoughts:

(3) Another option is to have the Blob API distinguish the two cases.
For example, "write()" always replaces all conents whilst "replace()"
always determines the true diffs.  This would expose some user control,
though it's perhaps a bit confusing!

(4) Even Array doesn't provide the ability to partially update it.
Somehow, this seems like a critical functionality as, even if diff's are
computed on writes, there will be situations where have the diff already
computed for us (e.g. it comes from the editor directly).

(5) Looking in Primitive.ByteArrayLayout.write(), it actually does
something pretty dumb.  It reads the whole incoming array into bytes and
writes them directly.  In fact, it could look and see whether the array
is a diff on the current blob, or not.  If the latter then no need to
write everything, only the diff.  Likewise, for other proxies.